cache_layered {cachem}	R Documentation

Compose any number of cache objects into a new, layered cache object

Description

Note that cache_layered is currently experimental.

Usage

cache_layered(..., logfile = NULL)

Arguments

...
Cache objects to compose into a new, layered cache object.

logfile
An optional filename or connection object to where logging information will be written. To log to the console, use stderr() or stdout().

Value

A layered caching object, with class cache_layered.

Examples

# Make a layered cache from a small memory cache and large disk cache
m <- cache_mem(max_size = 100 * 1024^2)
d <- cache_disk(max_size = 2 * 1024^3)
cl <- cache_layered(m, d)

[Package cachem version 1.0.6 Index]